% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeAssPlot.R
\name{AgeAssPlot}
\alias{AgeAssPlot}
\title{Function to plot age-structured models
Used within the \code{\link{GeMS}} function, not to be run on its own.}
\usage{
AgeAssPlot(REP, CTL, DAT, TRU, data2, MSEdir, CTLNameList)
}
\arguments{
\item{REP}{Read SimAss.rep file}

\item{CTL}{Read SimAss.CTL file}

\item{DAT}{Read SimAss.DAT file}

\item{TRU}{Read TrueQuantities.DAT file}

\item{data2}{Read Simass.PAR file}

\item{MSEdir}{Directory containing CTL files}

\item{CTLNameList}{Vector of CTL file names}
}
\value{
GeMS Object # To be implemented
}
\description{
Function to plot age-structured models
Used within the \code{\link{GeMS}} function, not to be run on its own.
}
