% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeStructureComp.R
\name{AgeStructureComp}
\alias{AgeStructureComp}
\title{Age-Structure Comparison}
\usage{
AgeStructureComp(Inout, RetroPeels = 6, CTLNames, MSEdir)
}
\arguments{
\item{RetroPeels}{Number of years to peel for retrospective analysis}

\item{MSEdir}{Directory containing CTL files}

\item{out}{List output from \code{\link{ReadCTLfile}}}

\item{CTLNameList}{Vector of CTL file names}

\item{plotNames}{Vector of the same length as CTLNameList with "prettier" names for plotting}
}
\value{
Plots comparing age-structured estimating models
}
\description{
Compare and plot age-structured models

Used within the \code{\link{run_GeMS}} function, can also be run on its own.
}
\examples{
\dontrun{
MSEdir <- "~/GeneralMSE/Examples/Cod_5_AgeStructure"
OMNames <- c("Cod_AgeStructure_CTL","Cod_Age_Mvary_CTL","Cod_Age_Mvary_estM_CTL")
out <- ReadCTLfile(OMNames[1])
AgeStructureComp(out=out,
                 CTLNameList=OMNames,
                 MSEdir=MSEdir,
                 plotNames=c("Base","Fixed M","Estimate M"))
}

}
