% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBHrec.R
\name{calcBHrec}
\alias{calcBHrec}
\title{Function that calculates recruitment values with Beverton-Holt function. For use in \code{\link{calcREF}}
Some overlap with \code{\link{BevHolt}}}
\usage{
calcBHrec(SRouts, spbioIn)
}
\arguments{
\item{SRouts}{Stock-recruitment parameter outputs}

\item{spbioIn}{Spawning stock biomass}
}
\value{
Negative log-likelihood
}
\description{
Function that calculates recruitment values with Beverton-Holt function. For use in \code{\link{calcREF}}
Some overlap with \code{\link{BevHolt}}
}
\examples{
\dontrun{

}

}
