% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcREF.R
\name{calcREF}
\alias{calcREF}
\title{Calculates... STUFF for use in Cody's brain. Probably related to \code{\link{StockRecCheck}}}
\usage{
calcREF(SRouts)
}
\arguments{
\item{SRouts}{List of something or other.}
}
\value{
STUFF
}
\description{
Calculates... STUFF for use in Cody's brain. Probably related to \code{\link{StockRecCheck}}
}
\examples{
\dontrun{
calcREF(DERP)
}
}
