% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRICKrec.R
\name{calcRICKrec}
\alias{calcRICKrec}
\title{Function that calculates recruitment values from Ricker function. For use in \code{\link{calcREF}}}
\usage{
calcRICKrec(SRouts, spbioIn)
}
\arguments{
\item{SRouts}{Stock-recruitment parameter outputs}

\item{spbioIn}{Spawning stock biomass}
}
\value{
Negative log-likelihood
}
\description{
Function that calculates recruitment values from Ricker function. For use in \code{\link{calcREF}}
}
\examples{
\dontrun{

}

}
