% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBias.R
\name{CalculateBias}
\alias{CalculateBias}
\title{Calculates bias in spawning stock biomass estimates for use in \code{\link{AgeStructureComp}}}
\usage{
CalculateBias(RetroOuts)
}
\arguments{
\item{RetroOuts}{Output from \code{\link{CheckRetro}}}
}
\value{
Matrix of biases
}
\description{
Calculates bias in spawning stock biomass estimates for use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{

}
}
