% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMohn.R
\name{CalculateMohn}
\alias{CalculateMohn}
\title{Calculates Mohn's Rho for use in \code{\link{AgeStructureComp}}}
\usage{
CalculateMohn(RetroOuts)
}
\arguments{
\item{RetroOuts}{Output from \code{\link{CheckRetro}}}
}
\value{
Matrix of Mohn's Rhos
}
\description{
Calculates Mohn's Rho for use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{

}
}
