% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckGradient.R
\name{CheckGradient}
\alias{CheckGradient}
\title{Extracts gradients from age-structured model runs for use in \code{\link{AgeStructureComp}}}
\usage{
CheckGradient(DrawDir, out, MSEdir)
}
\arguments{
\item{out}{List from \code{\link{ReadCTLfile}}}

\item{MSEdir}{Directory containing CTL file}

\item{CTLName}{Name of CTL file}
}
\value{
Matrix of gradients
}
\description{
Extracts gradients from age-structured model runs for use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
MSEdir <- "~/GeMS/Examples/Cod_1_Production"
CTLName <- "Cod_Base_CTL"
out <- ReadCTLfile(CTLName)
CheckGradient(CTLName,out,MSEdir)
}
}
