% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckReferencePoints.R
\name{CheckReferencePoints}
\alias{CheckReferencePoints}
\title{Reference Points
Extracts reference points from age-structured model runs for use in \code{\link{AgeStructureComp}}}
\usage{
CheckReferencePoints(DrawDir, out, MSEdir)
}
\arguments{
\item{out}{List from \code{\link{ReadCTLfile}}}

\item{CTLName}{Name of CTL file}

\item{MSEDir}{Directory containing CTL file}
}
\value{
List of reference points list(B35,F35,OFL,tB35,tF35,tOFL)
}
\description{
Reference Points
Extracts reference points from age-structured model runs for use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
MSEdir <- "~/GeneralMSE/Examples/Cod_5_AgeStructure"
OMName <- "Cod_AgeStructure_CTL"
out <- ReadCTLfile(OMName)
CheckReferencePoints(OMName, out, MSEdir)
}
}
