% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckRetro.R
\name{CheckRetro}
\alias{CheckRetro}
\title{Check Retrospective}
\usage{
CheckRetro(RetroPeels = 6, DrawDir, PlotRetro = 0, out, MSEdir)
}
\arguments{
\item{RetroPeels}{Number of years to conduct retrospective analysis}

\item{PlotRetro}{1/0 value to plot retrospective analyses}

\item{out}{List from \code{\link{ReadCTLfile}}}

\item{CTLName}{Name of CTL file}

\item{MSEDir}{Directory containing CTL file}
}
\value{
Predicted spawning biomass and true spawning biomass from retrospective analysis.
}
\description{
Extracts calculated retrospective bias from age-structured model runs for use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
CTLName <- "Cod_Base_CTL"
out <- ReadCTLfile(CTLName)
MSEdir <- "~/GeMS/Examples/Cod_1_Production"
CheckRetro(CTLName,out,MSEdir)
}
}
