% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanInput.R
\name{CleanInput}
\alias{CleanInput}
\title{Clean up OM input (checks lengths of vectors for time-varying input) for use in \code{\link{GeMS}}}
\usage{
CleanInput(input, SimYear)
}
\arguments{
\item{input}{Vector or value to be "cleaned"}

\item{SimYear}{Number of years in the simulation}
}
\value{
Vector of the same length or greater than SimYear
}
\description{
Clean up OM input (checks lengths of vectors for time-varying input) for use in \code{\link{GeMS}}
}
\examples{
\dontrun{
CheckReferencePoints("Cod_Base_CTL",ReadCTLfile("Cod_Base_CTL.csv"),"~/GeMS/Examples/Cod_1_Production")
}
}
