% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindF35.R
\name{FindF35}
\alias{FindF35}
\title{Calculates F35 for \code{\link{GeMS}}}
\usage{
FindF35(MaxAge, vulnN, vulnS, NatMn, NatMs, matureN, matureS, WeightAtAgeN,
  WeightAtAgeS, steepnessN, steepnessS, RzeroN, RzeroS, LenAtAgeN, LenAtAgeS,
  inRec, RefYear, fmortSin = 0, MovementN, MovementS)
}
\arguments{
\item{MaxAge}{Maximum age}

\item{vulnN}{Vulnerability matrix for northern population}

\item{vulnS}{Vulnerability matrix for southern population}

\item{NatMn}{Natural Mortality for northern population}

\item{NatMs}{Natural Mortality for southern population}

\item{matureN}{Proportions mature for northern population}

\item{matureS}{Proportions mature for southern population}

\item{WeightAtAgeN}{Weight-at-age matrix for northern population}

\item{WeightAtAgeS}{Weight-at-age matrix for southern population}

\item{steepnessN}{Steepness parameter for northern population}

\item{steepnessS}{Steepness parameter for southern population}

\item{RzeroN}{Virgin recruitment parameter for northern population}

\item{RzeroS}{Virgin recruitment parameter for southern population}

\item{LenAtAgeN}{Matrix of lengths at age for northern population}

\item{LenAtAgeS}{Matrix of lengths at age for southern population}

\item{inRec}{Historical recruitment values}

\item{RefYear}{Year of forecast}

\item{fmortSin}{Historical fishing mortality}

\item{MovementN}{Proportion of emigration to the southern population}

\item{MovementS}{Proportion of immigration into the northern population}
}
\value{
Vector of historical and calculated future Fs.
}
\description{
Calculates F35 for \code{\link{GeMS}}
}
