% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindFMSY.R
\name{FindFMSY}
\alias{FindFMSY}
\title{Calculates yield at F, to be passed to optimizing function}
\usage{
FindFMSY(x, MaxAge, VirInitN, VirInitS, vulnN, vulnS, NatMn, NatMs, matureN,
  matureS, WeightAtAgeN, WeightAtAgeS, steepnessN, steepnessS, RzeroN, RzeroS,
  LenAtAgeN, LenAtAgeS, sigmaRn, sigmaRs, RefYear, inputFs = 0, MovementN,
  MovementS)
}
\arguments{
\item{x}{Historical fishing mortality for nothern population}

\item{MaxAge}{Maximum age}

\item{VirInitN}{Initial biomass of northern population}

\item{VirInitS}{Initial biomass of southern population}

\item{vulnN}{Vulnerability matrix for northern population}

\item{vulnS}{Vulnerability matrix for southern population}

\item{NatMn}{Natural Mortality for northern population}

\item{NatMs}{Natural Mortality for southern population}

\item{matureN}{Proportions mature for northern population}

\item{matureS}{Proportions mature for southern population}

\item{WeightAtAgeN}{Weight-at-age matrix for northern population}

\item{WeightAtAgeS}{Weight-at-age matrix for southern population}

\item{steepnessN}{Steepness parameter for northern population}

\item{steepnessS}{Steepness parameter for southern population}

\item{RzeroN}{Virgin recruitment parameter for northern population}

\item{RzeroS}{Virgin recruitment parameter for southern population}

\item{LenAtAgeN}{Matrix of lengths at age for northern population}

\item{LenAtAgeS}{Matrix of lengths at age for southern population}

\item{sigmaRn}{Recruitment variability for northern population}

\item{sigmaRs}{Recruitment variability for southern population}

\item{RefYear}{Year of forecast}

\item{inputFs}{Historical fishing mortality for southern population}

\item{MovementN}{Proportion of emigration to the southern population}

\item{MovementS}{Proportion of immigration into the northern population}
}
\value{
Value of yield at F35%
}
\description{
Calculates yield at F, to be passed to optimizing function
}
\examples{
\dontrun{

}
}
