% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{GeMS}
\alias{GeMS}
\title{Wrapper function to run an MSE.}
\usage{
GeMS(out, CTLName, MSEdir = getwd(), silent = F, ADoptions = NA,
  ADsilent = T, echo = F, ...)
}
\arguments{
\item{out}{Named list output from \code{\link{ReadCTLfile}}}

\item{CTLName}{Name of CTL file}

\item{MSEdir}{Directory containing CTL files. Full absolute path preferred.}

\item{silent}{Logical; Showing output on console}

\item{ADoptions}{String of ADMB options to be passed. Always includes \code{-nohess}}

\item{ADsilent}{Logical; Showing ADMB output on console}

\item{echo}{Logical; Write to an echo file to show quantities from OM}

\item{...}{WHO KNOWS}
}
\value{
GeMS Object # To be implemented
}
\description{
Function that runs the entire MSE- generates data from OM, assesses it with EM, and forecasts. Can be run on its own, but it is intended to be run with \code{\link{run_GeMS}}
}
\examples{
\dontrun{
GeMS(out=ReadCTLfile("Cod_LowProd_CTL.csv"),CTLName="Cod_LowProd_CTL",MSEdir="~/Cod_1_Produtction",ADoptions="-gbs 1000000",ADsilent=T)
}
}
