% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMSplots.R
\name{GeMSplots}
\alias{GeMSplots}
\title{More plots.. currently commented out from \code{\link{AgeStructureComp}}}
\usage{
GeMSplots(SourceFolder, out)
}
\arguments{
\item{out}{List output from \code{\link{ReadCTLfile}}}

\item{CTLName}{Name of CTL file}

\item{MSEdir}{Directory containing CTL files}
}
\value{
The path to an GeMS Age-Structured Model binary.
}
\description{
More plots.. currently commented out from \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
GeMSplots("Cod_Base_CTL",ReadCTLfile("Cod_Base_CTL.csv"),"~/GeMS/Examples/Cod_1_Production")
}

}
