% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gems.r
\name{generateParameterMatrix}
\alias{generateParameterMatrix}
\title{generate a template for mean parameters}
\usage{
generateParameterMatrix(hf)
}
\arguments{
\item{hf}{a \code{transition.structure} of dimension \eqn{N \times N}{N x
N}, where \eqn{N} is the number of states.}
}
\value{
a \code{transition.structure} of dimension \eqn{N \times N}{N x N},
whose components are \code{lists} of the right length for the parameters in
the corresponding hazard function \code{hf}.
}
\description{
This function simplifies generating the matrix of mean parameters from a
matrix of transition functions.
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\references{
Nello Blaser, Luisa Salazar Vizcaya, Janne Estill, Cindy Zahnd, 
Bindu Kalesan, Matthias Egger, Olivia Keiser, Thomas Gsponer (2015). gems: 
An R Package for Simulating from Disease Progression Models. Journal of 
Statistical Software, 64(10), 1-22. URL http://www.jstatsoft.org/v64/i10/.
}
\seealso{
\code{\link{transition.structure}},
\code{\link{simulateCohort}}
}
\keyword{utilities}

