% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exec.R
\name{get_exec}
\alias{get_exec}
\title{Get Executable
Get SimAss binary/executable location in package}
\usage{
get_exec(exe_name = SimAssExec)
}
\arguments{
\item{exe_name}{Name of the SimAss binary}
}
\value{
The path to a GeMS Age-Structured Model binary.
}
\description{
Get Executable
Get SimAss binary/executable location in package
}
\examples{
\dontrun{
get_exec("simass.exe")
}
}
