% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestControlRule.R
\name{HarvestControlRule}
\alias{HarvestControlRule}
\title{Harvest Control Rule
Used within the \code{\link{GeMS}} function.}
\usage{
HarvestControlRule(FMSY, BMSY, ExploitBio, SpawnBio, alpha, beta,
  HarvestControl, ConstantCatch = 0, ConstantF = 0)
}
\arguments{
\item{FMSY}{Fishing mortality at MSY}

\item{BMSY}{Biomass at MSY}

\item{ExploitBio}{Exploitable biomass}

\item{SpawnBio}{Spawning biomass}

\item{alpha}{Currently not implemented}

\item{beta}{Slope of harvest control rule}

\item{HarvestControl}{Harvest control rule (options in CTL file)}

\item{ConstantCatch}{Currently not implemented}

\item{ConstantF}{Currently not implemented}

\item{RetroPeels}{Number of years to peel for retrospective analysis}

\item{CTLNameList}{Vector of CTL file names}

\item{MSEdir}{Directory containing CTL files}

\item{plotNames}{Vector of the same length as CTLNameList with "prettier" names for plotting}
}
\value{
Value of Total Allowable Catch
}
\description{
Harvest Control Rule
Used within the \code{\link{GeMS}} function.
}
