% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialN.R
\name{initialN}
\alias{initialN}
\title{Calculate initial biomass using \eqn{Recruitment * exp(-NaturalMortality)}}
\usage{
initialN(Rzero, NatM, inAge)
}
\arguments{
\item{Rzero}{Virgin recruitment}

\item{NatM}{Natural mortality}

\item{inAge}{Maximum age}
}
\value{
Vector of initial biomass
}
\description{
Used in \code{\link{ProjPopDym}}, \code{\link{FindF35}}, \code{\link{SBPRfunc}}
}
