% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyBoxPlot.R
\name{MyBoxPlot}
\alias{MyBoxPlot}
\title{Customised Box Plot
Used within the \code{\link{GeMSplots}} function, not quite implemented.}
\usage{
MyBoxPlot(Inputs, Title, bottomplot = 0, refVal = 0)
}
\arguments{
\item{Inputs}{Matrix of values to be plotted}

\item{Title}{Title of the plot}

\item{bottomplot}{1/0 if figure is at the bottom of plot, includes the legend}

\item{refVal}{Reference value to set maximum y limit}
}
\value{
Boxplots
}
\description{
Customised Box Plot
Used within the \code{\link{GeMSplots}} function, not quite implemented.
}
\examples{
\dontrun{
MyBoxPlot(FMSY[,InitYear:SimYear],"FMSY")
}

}
