% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stuff_inside.R
\name{plot_stuff_inside}
\alias{plot_stuff_inside}
\title{Compare model estimates
Used within the \code{\link{AgeStructureComp}} function to compare true values with estimated values.}
\usage{
plot_stuff_inside(input1, input2, title, CTLNameList, plotNames = NA,
  nSim = Nsim)
}
\arguments{
\item{input1}{Array of true quantities}

\item{input2}{Array of estimated quantities}

\item{title}{Quantity being plotted, e.g. "Catch", "Survey Index", etc.}

\item{CTLNameList}{Vector of CTL file names}

\item{plotNames}{Vector of the same length as CTLNameList with "prettier" names for plotting}

\item{nSim}{Number of replicates per OM, specified Nsim in CTL file}
}
\value{
Plot comparing age-structured estimating models
}
\description{
Compare model estimates
Used within the \code{\link{AgeStructureComp}} function to compare true values with estimated values.
}
