% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLifeHistory.R
\name{PlotLifeHistory}
\alias{PlotLifeHistory}
\title{Plot Life History}
\usage{
PlotLifeHistory(LenAtAgeN, LenAtAgeS, matureN, matureS, vulnN, vulnS, survSelN,
  survSelS, WeightAtAgeN, WeightAtAgeS, MovementN, MovementS, NatMs, NatMn,
  VirBioN, VirBioS, RzeroN, RecErrN, steepnessN, steepnessS, RzeroS, RecErrS,
  sigmaRn, sigmaRs, HistoricalFn, HistoricalFs, SimYear, MaxAge)
}
\arguments{
\item{LenAtAgeN}{Matrix of lengths at age for northern population}

\item{LenAtAgeS}{Matrix of lengths at age for southern population}

\item{matureN}{Proportions mature for northern population}

\item{matureS}{Proportions mature for southern population}

\item{vulnN}{Vulnerability matrix for northern population}

\item{vulnS}{Vulnerability matrix for southern population}

\item{survSelN}{Survey selectivity matrix for northern population}

\item{survSelS}{Survey selectivity matrix for southern population}

\item{WeightAtAgeN}{Weight-at-age matrix for northern population}

\item{WeightAtAgeS}{Weight-at-age matrix for southern population}

\item{MovementN}{Proportion of emigration to the southern population}

\item{MovementS}{Proportion of immigration into the northern population}

\item{NatMs}{Natural Mortality for southern population}

\item{NatMn}{Natural Mortality for northern population}

\item{VirBioN}{Initial virgin biomass for northern population}

\item{VirBioS}{Initial virgin biomass for southern population}

\item{RzeroN}{Virgin recruitment parameter for northern population}

\item{steepnessN}{Steepness parameter for northern population}

\item{steepnessS}{Steepness parameter for southern population}

\item{RzeroS}{Virgin recruitment parameter for southern population}

\item{RecErrS}{Recruitment deviations}

\item{sigmaRn}{Recruitment variability for northern population}

\item{sigmaRs}{Recruitment variability for southern population}

\item{HistoricalFn}{Vector of historical fishing mortalities for northern population}

\item{HistoricalFs}{Vector of historical fishing mortalities for southern population}

\item{SimYear}{Year of simulation}

\item{MaxAge}{Maximum age}
}
\value{
Plots plots plots
}
\description{
Plots life history based on inputs from CTL file, used in \code{\link{GeMS}}
}
