% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotParam.R
\name{PlotParam}
\alias{PlotParam}
\title{Plot Life History Parameter}
\usage{
PlotParam(param, OMfile, EMvec, plotnames = NA, dolegend = F,
  ylims = NA, ...)
}
\arguments{
\item{param}{Parameter of interest, named from PAR or CTL file (one of c("NatM","GrowthK","Linf","SelPars50","SelPars95","NatMn","VonKn","LinfN","sel50n","sel95n"))}

\item{OMfile}{Name of CTL file to base truth on}

\item{EMvec}{Vector of EM names}

\item{dolegend}{Logical; include legend?}

\item{ylims}{vector of length two containing limits for y-axis}

\item{...}{any other graphical parameters, e.g., ylab,xlab}
}
\value{
Plots plots plots; list of truth and estimates pulled from PAR and CTL files.
}
\description{
Plots life history parameter estimates from final assessment year
}
