% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonPlots.R
\name{PolygonPlots}
\alias{PolygonPlots}
\title{Plots polygons}
\usage{
PolygonPlots(Truth = NA, Estimated = NA, Observed = NA,
  AddLegend = F, bottom = F, quantity = NA, SimYear = NA,
  Nsim = NA, ylimIN = NA)
}
\value{
Plots plots plots
}
\description{
Plots polygons
}
