% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProdMod.R
\name{ProdMod}
\alias{ProdMod}
\title{Production Model
Fits a production model to input data}
\usage{
ProdMod(x, CatchData, IndexData, estInit = 0)
}
\arguments{
\item{x}{Vector of production model parameters, c(carrying capacity, population growth rate, initial biomass as a proportion of carrying capacity)}

\item{CatchData}{Vector of catch}

\item{IndexData}{Vector of abundance indices}

\item{EstInit}{1/0 Estimate initial biomass as proportion}
}
\value{
Sum Squared Errors from fitting production model
}
\description{
Production Model
Fits a production model to input data
}
\examples{
\dontrun{
K <- 10000
r <- 0.4
# Not needed if not estimating initial biomass
InitBio <- 0.95
x <- c(K,r,InitBio)
Catch <- c(500, 900, 750)
Index <- c(0.5, 0.9, 0.75)
ProdMod(x, Catch, Index, estInit=1)
}

}
