% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProdModPlot.R
\name{ProdModPlot}
\alias{ProdModPlot}
\title{Production Model Plot}
\usage{
ProdModPlot(x, CatchData, IndexData, plots = 0, estInit = 0)
}
\arguments{
\item{x}{Vector of production model parameters, c(carrying capacity, population growth rate, initial biomass as a proportion of carrying capacity)}

\item{CatchData}{Vector of catch}

\item{IndexData}{Vector of abundance indices}

\item{plots}{1/0 To plot, or not to plot}

\item{EstInit}{1/0 Estimate initial biomass as proportion}
}
\value{
Plot and vector of predicted biomass.
}
\description{
Production Model Plot
}
\examples{
\dontrun{
K <- log(10000)
r <- 0.4
# Not needed if not estimating initial biomass
InitBio <- 0.95
x <- c(K,r,InitBio)
Catch <- c(500, 900, 750)
Index <- c(0.5, 0.9, 0.75)
ProdModPlot(x, Catch, Index, estInit=1)
}

}
