% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductionModelOutput.R
\name{ProductionModelOutput}
\alias{ProductionModelOutput}
\title{Compare Production Model Outputs}
\usage{
ProductionModelOutput(out, CTLNameList, MSEdir, ylimIN = NA, plotNames = NA)
}
\arguments{
\item{out}{List ouput from \code{\link{ReadCTLfile}}}

\item{CTLNameList}{List of CTL files}

\item{MSEdir}{Directory containing CTL files}

\item{ylimIN}{Vector of optional manual limits for y-axis}

\item{plotNames}{Vector of the same length as CTLNameList with "prettier" names for plotting}
}
\value{
GeMS Object # To be implemented
}
\description{
Analyse and compare output from production estimating models
}
\examples{
\dontrun{
MSEdir <- "~/GeneralMSE/Examples/Cod_1_Production"
OMNames <- c("Cod_LowProd_CTL","Cod_Base_CTL","Cod_HighProd_CTL")
out <- ReadCTLfile(OMNames[1])
ProductionModelOutput(out,OMNames,MSEdir,plotNames=c("Low Productivity", "Base", "High Productivity"))
}
}
