% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjPopDym.R
\name{ProjPopDym}
\alias{ProjPopDym}
\title{Project population dynamics}
\usage{
ProjPopDym(fmortN, fmortS = 0, MaxAge, vulnN, vulnS, NatMn, NatMs,
  matureN, matureS, WeightAtAgeN, WeightAtAgeS, steepnessN, steepnessS,
  RzeroN, RzeroS, LenAtAgeN, LenAtAgeS, ProxyRec = 0, sigmaRn, sigmaRs,
  RefYear, MovementN, MovementS)
}
\value{
List of projected yield and biomass
}
\description{
Project population dynamics
}
