% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullTimevary.R
\name{PullTimevary}
\alias{PullTimevary}
\title{Used to pull parameters from age-structured estimation model output. For use in \code{\link{AgeStructureComp}}}
\usage{
PullTimevary(CTLNames, out, MSEdir)
}
\arguments{
\item{out}{List output from \code{\link{ReadCTLfile}}}

\item{MSEdir}{Directory containing CTL files}

\item{CTLNameList}{Vector of CTL files}
}
\value{
List of quantities
}
\description{
Used to pull parameters from age-structured estimation model output. For use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
MSEdir <- "~/GeneralMSE/Examples/Cod_5_AgeStructure"
OMNames <- c("Cod_AgeStructure_CTL","Cod_Age_Mvary_CTL","Cod_Age_Mvary_estM_CTL")
out <- ReadCTLfile(OMNames[1])
AgeStructureComp(out=out,
                 CTLNameList=OMNames,
                 MSEdir=MSEdir,
                 plotNames=c("Base","Fixed M","Estimate M"))
}

}
