% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullTrueProj.R
\name{PullTrueProj}
\alias{PullTrueProj}
\title{Pull True Projections
Not used anywhere... yet.}
\usage{
PullTrueProj(CTLNameList, out, MSEdir)
}
\arguments{
\item{CTLNameList}{Vector of CTL file names}

\item{out}{List output from \code{\link{ReadCTLfile}}}

\item{MSEdir}{Directory containing CTL files}
}
\value{
Plots comparing age-structured estimating models
}
\description{
Pull True Projections
Not used anywhere... yet.
}
\examples{
\dontrun{
MSEdir <- "~/GeneralMSE/Examples/Cod_5_AgeStructure"
OMNames <- c("Cod_AgeStructure_CTL","Cod_Age_Mvary_CTL","Cod_Age_Mvary_estM_CTL")
out <- ReadCTLfile(OMNames[1])
AgeStructureComp(out=out,
                 CTLNameList=OMNames,
                 MSEdir=MSEdir,
                 plotNames=c("Base","Fixed M","Estimate M"))
}

}
