% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCTLfile.R
\name{ReadCTLfile}
\alias{ReadCTLfile}
\title{Read in CTL csv file}
\usage{
ReadCTLfile(input)
}
\arguments{
\item{input}{Name of csv file}
}
\value{
List of inputs from CTL file
}
\description{
Read in CTL csv file
}
\examples{
\dontrun{
ReadCTLfile("Cod_Base_CTL")

ReadCTLfile("~/GeMS/Examples/Cod_1_Production/Cod_Base_CTL")
}

}
