% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPar.R
\name{ReadPar}
\alias{ReadPar}
\title{Pull parameters from .PAR file. For use in \code{\link{AgeStructureComp}}}
\usage{
ReadPar(wd)
}
\arguments{
\item{wd}{Directory containing .PAR file from running age-structured model.}
}
\value{
Named list of quantities from .PAR file
}
\description{
Pull parameters from .PAR file. For use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
WorkingDir <- "~/GeMS/inst/exdata/Cod_5_AgeStructure/Cod_AgeStructure_CTL/1/70"
ReadPar(WorkingDir)
}

}
