% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRep.R
\name{ReadRep}
\alias{ReadRep}
\title{Pull parameters from .REP file. For use in \code{\link{AgeStructureComp}}}
\usage{
ReadRep(wd)
}
\arguments{
\item{wd}{Directory containing simass.rep file from running age-structured model.}
}
\value{
Named list of quantities from .REP file
}
\description{
Pull parameters from .REP file. For use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
WorkingDir <- "~/GeMS/inst/exdata/Cod_5_AgeStructure/Cod_AgeStructure_CTL/1/70"
ReadPar(WorkingDir)
}

}
