% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTruDat.R
\name{ReadTruDat}
\alias{ReadTruDat}
\title{Pull parameters from TrueQuantities.DAT file. For use in \code{\link{AgeStructureComp}}}
\usage{
ReadTruDat(wd)
}
\arguments{
\item{wd}{Directory containing TrueQuantities.DAT file from running age-structured model.}
}
\value{
Named list of quantities from TrueQuantities.DAT file
}
\description{
Pull parameters from TrueQuantities.DAT file. For use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
WorkingDir <- "~/GeMS/inst/exdata/Cod_5_AgeStructure/Cod_AgeStructure_CTL/1/70"
ReadTruDat(WorkingDir)
}

}
