% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruitment.R
\name{Recruitment}
\alias{Recruitment}
\title{Recruitment function}
\usage{
Recruitment(EggsIN, steepnessIN, RzeroIN, RecErrIN, recType, NatMin,
  vulnIN, matureIN, weightIN, LenAtAgeIN, MaxAge, sigmaRin)
}
\arguments{
\item{EggsIN}{Vector of egg mass}

\item{steepnessIN}{Steepness parameter}

\item{RzeroIN}{Virgin recruitment}

\item{RecErrIN}{Recruitment error}

\item{recType}{"BH" or "Rick"; For Beverton-Holt or Ricker stock-recruitment relationship}

\item{NatMin}{Natural mortality}

\item{vulnIN}{Vector of vulnerability-at-age}

\item{matureIN}{Vector of maturity-at-age}

\item{weightIN}{Vector of weight-at-age}

\item{LenAtAgeIN}{Vector of lengths-at-age}

\item{MaxAge}{Maximum age}

\item{sigmaRin}{Recruitment variability}
}
\value{
Vector of recruitment values
}
\description{
Applies selected stock-recruitment function using life history parameters. Used in \code{\link{GeMS}}, \code{\link{PlotLifeHistory}}, and \code{\link{ProjPopDym}}.
}
