% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GeMS.R
\name{run_GeMS}
\alias{run_GeMS}
\title{Function to run GeMS}
\usage{
run_GeMS(CTLNameList, MSEdir = getwd(), runparallel = F, cores = 1,
  silent = T, ...)
}
\arguments{
\item{CTLNameList}{Vector of CTL files}

\item{MSEdir}{Directory containing CTL files}

\item{runparallel}{Logical; operating models should be run in parallel. If you're doing this, note that the package \pkg{foreach} will be loaded.}

\item{cores}{Number of cores to be used for parallel runs}

\item{silent}{Logical; Show output on console}

\item{...}{Anything to be passed to \code{\link{GeMS}} or \code{\link{AgeStructureComp}} or \code{\link{ProductionModelOutput}}. Can include \code{plotNames}, \code{ADoptions} and \code{ADsilent} for ADMB options.}
}
\value{
\code{\link{GeMS}} output
}
\description{
Main function to run a full Generalized MSE.
}
\examples{
\dontrun{
OMNames <- c("Cod_LowProd_CTL","Cod_Base_CTL","Cod_HighProd_CTL")
MSEdir <- "~/GeneralMSE/Examples/Cod_1_Production"
run_GeMS(OMNames, MSEdir)

# or
run_GeMS(OMNames, MSEdir, runparallel=T, cores=2, plotNames = c("Low Productivity", "Base", "High Productivity"))
}
}
\seealso{
\code{\link{GeMS}}
}
