% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBPRfunc.R
\name{SBPRfunc}
\alias{SBPRfunc}
\title{Calculate spawning-biomass-per-recruit quantities for use in \code{\link{Recruitment}}}
\usage{
SBPRfunc(F, Rzero, NatM, vuln, mature, weight, MaxAge)
}
\arguments{
\item{F}{Fishing mortality}

\item{Rzero}{Virgin recruitment}

\item{NatM}{Natural mortality}

\item{vuln}{Vector of vulnerability-at-age}

\item{mature}{Vector of maturity-at-age}

\item{weight}{Vector of weight-at-age}

\item{MaxAge}{Maximum age}
}
\value{
List of mature spawning biomass (SBP), spawning biomass per recruit (SBPR), and yield per recruit (YPR).
}
\description{
Calculate spawning-biomass-per-recruit quantities for use in \code{\link{Recruitment}}
}
