% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockRecCheck.R
\name{StockRecCheck}
\alias{StockRecCheck}
\title{Stock-Recruitment Check}
\usage{
StockRecCheck(rec, spbio, plotSR = F)
}
\arguments{
\item{rec}{Vector of recruitment values}

\item{spbio}{Vector of spawning biomass values (must be of same length as \code{rec})}

\item{plotSR}{Logical; Plot stock-recruitment relationship}
}
\value{
Recruitment function to be used
}
\description{
Pre-MSE, used to determine stock-recruitment relationship (Average, Beverton-Holt or Ricker), likely for use somewhere eventually. Currently uses \code{SRstarts} function that has not previously been specified. Probably related to \code{\link{calcREF}}
}
