% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrLine.R
\name{StrLine}
\alias{StrLine}
\title{Negative log-likelihood}
\usage{
StrLine(x, rec)
}
\arguments{
\item{x}{Vector of length two, c(predicted value and sigma)}

\item{rec}{Vector of recruitment values}
}
\value{
Negative log-likelihood value
}
\description{
Used in \code{\link{StockRecCheck}}.
}
