% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TakeOut.R
\name{TakeOut}
\alias{TakeOut}
\title{Extract values from given string for use in \code{\link{ReadCTLfile}}. Uses \code{grep} function.}
\usage{
TakeOut(SearchTerm, SearchPool)
}
\arguments{
\item{SearchTerm}{Term to be searched}

\item{SearchPool}{String to be searched over}
}
\value{
Vector of numbers or characters.
}
\description{
Extract values from given string for use in \code{\link{ReadCTLfile}}. Uses \code{grep} function.
}
\examples{
\dontrun{
TakeOut(SearchTerm="Nsim",SearchPool=as.matrix(read.csv(OMName)))
}
}
