% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gems.r
\docType{class}
\name{transition.structure}
\alias{[[,transition.structure-method}
\alias{[[.transition.structure}
\alias{[[<-,transition.structure-method}
\alias{[[<-.transition.structure}
\alias{possibleTransitions}
\alias{possibleTransitions,transition.structure-method}
\alias{print,transition.structure-method}
\alias{print.transition.structure}
\alias{transition.structure}
\alias{transition.structure-class}
\title{Class \code{"transition.structure"}}
\usage{
\S4method{[[}{transition.structure}(x, i, j, ..., exact = TRUE)

\S4method{[[}{transition.structure}(x, i, j) <- value

possibleTransitions(object)

\S4method{possibleTransitions}{transition.structure}(object)

\S4method{print}{transition.structure}(x)
}
\arguments{
\item{x,object}{the \code{transition.structure}}

\item{i,j}{same as for \code{matrix}}

\item{exact,value,...}{passed on to \code{list} method}
}
\description{
This S4 class provides a structure to specify different characteristics of
transitions, such as transition functions functions, parameters or parameter
covariances.
}
\section{Slots}{

\describe{
\item{\code{states.number}}{Object of class \code{"numeric"}: number of states}

\item{\code{list.matrix}}{Object of class \code{"matrix"}: a \code{list} with two
dimensions, where list element [i,j] correspond to transitions from i to j}
}}
\section{Objects from the Class}{
 Objects are created by calls to the
functions \code{generateHazardMatrix}, \code{generateParameterMatrix},
\code{generateParameterCovarianceMatrix}.
}
\examples{
showClass("transition.structure")
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\seealso{
\code{\link{generateHazardMatrix}},
\code{\link{generateParameterMatrix}},
\code{\link{generateParameterCovarianceMatrix}}
}
\keyword{classes}

