% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gems-package.R
\docType{class}
\name{ArtCohort-class}
\alias{ArtCohort-class}
\alias{[,ArtCohort,ANY,ANY,ANY-method}
\alias{[,ArtCohort-method}
\alias{head,ArtCohort-method}
\alias{summary,ArtCohort-method}
\alias{tail,ArtCohort-method}
\alias{update,ArtCohort-method}
\title{Class \code{"ArtCohort"}}
\description{
Is a S4 class for the artificial cohort generated by \code{simulateCohort}.
}
\section{Objects from the Class}{
 Objects are created by calls to the
function \code{simulateCohort}.
}
\examples{
showClass("ArtCohort")
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\seealso{
\code{\link{simulateCohort}},
\code{\link{transition.structure-class}},
\code{\link{transitionProbabilities}}, \code{\link{cumulativeIncidence}}
}
\keyword{classes}

