% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gems-package.R
\docType{package}
\name{gems}
\alias{gems}
\alias{gems-package}
\title{gems: Generalized multistate simulation model}
\description{
The gems package allows to simulate and analyze multistate models with
general hazard functions. It provides functionality for the preparation
of hazard functions and parameters, simulation from a general multistate
model and predicting future events. The multistate model is not required
to be a Markov model and may take the history of previous events into
account. In the basic version, it allows to simulate from
transition-specific hazard function, whose parameters are multivariable
normally distributed.
}

