% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gems-package.R
\docType{package}
\name{gems}
\alias{gems}
\alias{gems-package}
\title{gems: Generalized Multistate Simulation Model}
\description{
Simulate and analyze multistate models with
general hazard functions. gems provides functionality for the preparation
of hazard functions and parameters, simulation from a general multistate
model and predicting future events. The multistate model is not required
to be a Markov model and may take the history of previous events into
account. In the basic version, it allows to simulate from
transition-specific hazard function, whose parameters are multivariable
normally distributed.
}
\references{
Nello Blaser, Luisa Salazar Vizcaya, Janne Estill, Cindy Zahnd,
Bindu Kalesan, Matthias Egger, Olivia Keiser, Thomas Gsponer (2015). gems:
An R Package for Simulating from Disease Progression Models. Journal of
Statistical Software, 64(10), 1-22. URL http://www.jstatsoft.org/v64/i10/.
}

