% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gems.r
\name{generateHazardMatrix}
\alias{generateHazardMatrix}
\title{generate template for transition functions}
\usage{
generateHazardMatrix(statesNumber)
}
\arguments{
\item{statesNumber}{the number of states to be considered.}
}
\value{
a \code{transition.structure} of dimension \eqn{N \times N}{N x N},
where \eqn{N} is the number of states and with value "impossible" for all
potential transitions.
}
\description{
This function simplifies generating the matrix of transition functions.
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\references{
Nello Blaser, Luisa Salazar Vizcaya, Janne Estill, Cindy Zahnd,
Bindu Kalesan, Matthias Egger, Olivia Keiser, Thomas Gsponer (2015). gems:
An R Package for Simulating from Disease Progression Models. Journal of
Statistical Software, 64(10), 1-22. URL http://www.jstatsoft.org/v64/i10/.
}
\seealso{
\code{\link{transition.structure}},
\code{\link{simulateCohort}}
}
\keyword{utilities}

