% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gems.r
\name{generateParameterCovarianceMatrix}
\alias{generateParameterCovarianceMatrix}
\title{generate a template for parameter covariances}
\usage{
generateParameterCovarianceMatrix(mu)
}
\arguments{
\item{mu}{a \code{transition.structure} of dimension \eqn{N \times N}{N x
N}, whose components \code{list} the mean values for the parameters in the
\code{transitionFunction}.}
}
\value{
a \code{transition.structure} of dimension \eqn{N \times N}{N x N}
of covariance matrices for the \code{parameters}.
}
\description{
This function simplifies generating the matrix of parameter covariances from
a matrix of mean parameters.
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\references{
Nello Blaser, Luisa Salazar Vizcaya, Janne Estill, Cindy Zahnd,
Bindu Kalesan, Matthias Egger, Olivia Keiser, Thomas Gsponer (2015). gems:
An R Package for Simulating from Disease Progression Models. Journal of
Statistical Software, 64(10), 1-22. URL http://www.jstatsoft.org/v64/i10/.
}
\seealso{
\code{\link{transition.structure}},
\code{\link{generateParameterMatrix}}, \code{\link{simulateCohort}}
}
\keyword{utilities}

