% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gems-package.R
\docType{class}
\name{PosteriorProbabilities-class}
\alias{PosteriorProbabilities-class}
\alias{[,PosteriorProbabilities,ANY,ANY,ANY-method}
\alias{[,PosteriorProbabilities-method}
\alias{head,PosteriorProbabilities-method}
\alias{plot,PosteriorProbabilities-method}
\alias{tail,PosteriorProbabilities-method}
\title{Class \code{"PosteriorProbabilities"}}
\description{
This S4 class summarizes the posterior probabilities over time for objects
of class \code{"ArtCohort"}
}
\section{Objects from the Class}{
 Objects are created by calls to the
function \code{simulateCohort}.
}
\examples{
showClass("PosteriorProbabilities")
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\seealso{
\code{\link{transitionProbabilities}},
\code{\link{cumulativeIncidence}}, \code{\link{ArtCohort-class}}
}
\keyword{classes}

