% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullTimevary.R
\name{PullTimevary}
\alias{PullTimevary}
\title{Used to pull parameters from age-structured estimation model output. For use in \code{\link{AgeStructureComp}}}
\usage{
PullTimevary(out, MSEdir)
}
\arguments{
\item{out}{List output from \code{\link{ReadCTLfile}}}

\item{MSEdir}{Directory containing CTL files}
}
\value{
List of quantities
}
\description{
Used to pull parameters from age-structured estimation model output. For use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
MSEdir <- "~/GeneralMSE/Examples/Cod_5_AgeStructure"
out <- ReadCTLfile("Cod_AgeStructure_CTL")
PullTimevary(out=out,
                 MSEdir=MSEdir)
}

}
