% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullTimevary.R
\name{PullTimevary}
\alias{PullTimevary}
\title{Used to pull parameters from age-structured estimation model output. For use in \code{\link{AgeStructureComp}}}
\usage{
PullTimevary(out, MSEdir, CTLNameList)
}
\arguments{
\item{out}{List output from \code{\link{ReadCTLfile}}}

\item{MSEdir}{Directory containing CTL files}

\item{CTLNameList}{Vector of CTL names}
}
\value{
List of quantities
}
\description{
Used to pull parameters from age-structured estimation model output. For use in \code{\link{AgeStructureComp}}
}
\examples{
\dontrun{
MSEdir <- "~/GeMS/inst/extdata/Cod_5_AgeStructure"
CTLName <- "Cod_Age_Structure_CTL"
out <- ReadCTLfile(CTLName)
PullTimevary(out=out,
             MSEdir=MSEdir,
             CTLNameList=CTLName)
}

}
