% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gems-package.R
\docType{data}
\name{tavi}
\alias{tavi}
\title{tavi data set}
\format{A data frame with 194 observations on the following 7 variables.
\describe{ \item{list("id")}{a character vector that contains the patient
id's} \item{list("kidney")}{a numeric vector; indicator variable that show
if an event has occurred} \item{list("kidney.dur")}{a numeric vector; times
at which the events occurred or the patients were censored}
\item{list("bleeding")}{a numeric vector; indicator variable that show if an
event has occurred} \item{list("bleeding.dur")}{a numeric vector; times at
which the events occurred or the patients were censored}
\item{list("death")}{a numeric vector; indicator variable that show if an
event has occurred} \item{list("death.dur")}{a numeric vector; times at
which the events occurred or the patients were censored} }}
\description{
The simulated data set for each patient contains data for kidney injuries,
bleeding complications and the combined endpoint of stroke or death. The
data was simulated from the original data following the steps described in
the package vignette.
}
\examples{
head(data(tavi))
}
\keyword{datasets}

