% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gems.r
\name{transitionProbabilities}
\alias{transitionProbabilities}
\title{transition probabilities}
\usage{
transitionProbabilities(object, times, M = 100, stateNames = paste("State",
  as.list(1:dim(cohorts)[1])))
}
\arguments{
\item{object}{either the output of \code{\link{simulateCohort}} or the
\code{matrix} with the \code{probabilities} slot of that output.}

\item{times}{a time vector.}

\item{M}{number of groups for calculating confidence intervals.}

\item{stateNames}{a list with the names of states.}
}
\value{
an object of class \code{"PosteriorProbabilities"}, containing the
statenames, timepoints and the transition probabilities with pointwise
prediction intervals over time.
}
\description{
Calculates the probabilities and prediction intervals after first state
}
\author{
Luisa Salazar Vizcaya, Nello Blaser, Thomas Gsponer
}
\references{
Nello Blaser, Luisa Salazar Vizcaya, Janne Estill, Cindy Zahnd,
Bindu Kalesan, Matthias Egger, Olivia Keiser, Thomas Gsponer (2015). gems:
An R Package for Simulating from Disease Progression Models. Journal of
Statistical Software, 64(10), 1-22. URL http://www.jstatsoft.org/v64/i10/.
}
\seealso{
\code{\link{PosteriorProbabilities}},
\code{\link{ArtCohort}}, \code{\link{simulateCohort}}
}
\keyword{utilities}

